chat-ui/requirements.txt
streamlit>=1.30.0
google-cloud-aiplatform[agent_engines]>=1.95.0
python-dotenv>=1.0.1
requests>=2.32.3
